+++ scilab/modules/graphics/src/c/getHandleProperty/set_alignment_property.c
@@ -25,7 +25,7 @@
 #include "localization.h"
 
 /*------------------------------------------------------------------------*/
-int set_alignment_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_alignment_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   if ( !isParameterStringMatrix( valueType ) )
