+++ scilab/modules/graphics/src/c/getHandleProperty/set_auto_resize_property.c
@@ -27,7 +27,7 @@
 #include "GraphicSynchronizerInterface.h"
 
 /*------------------------------------------------------------------------*/
-int set_auto_resize_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_auto_resize_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   int status = SET_PROPERTY_ERROR;
   if ( !isParameterStringMatrix( valueType ) )
