+++ scilab/modules/graphics/src/c/getHandleProperty/set_auto_ticks_property.c
@@ -29,7 +29,7 @@
 #include "localization.h"
 
 /*------------------------------------------------------------------------*/
-int set_auto_ticks_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_auto_ticks_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   char ** values = getStringMatrixFromStack( stackPointer ) ;
