+++ scilab/modules/graphics/src/c/getHandleProperty/set_callback_type_property.c
@@ -27,7 +27,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_callback_type_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_callback_type_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   int cbType = -1;
