+++ scilab/modules/graphics/src/c/getHandleProperty/set_data_property.c
@@ -630,7 +630,7 @@ int CheckAndUpdate_z_shift(sciPointObj * pobj, int numrow)
   return SET_PROPERTY_SUCCEED ;
 }
 /*------------------------------------------------------------------------*/
-int set_data_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_data_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   if( sciGetEntityType(pobj) == SCI_SEGS && pSEGS_FEATURE(pobj)->ptype == 1 )
   {
