+++ scilab/modules/graphics/src/c/getHandleProperty/set_dimension_property.c
@@ -26,7 +26,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_dimension_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_dimension_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   double * values = getDoubleMatrixFromStack( stackPointer ) ;
