+++ scilab/modules/graphics/src/c/getHandleProperty/set_figure_style_property.c
@@ -31,7 +31,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_figure_style_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_figure_style_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   if ( !isParameterStringMatrix( valueType ) )
