+++ scilab/modules/graphics/src/c/getHandleProperty/set_font_angle_property.c
@@ -27,7 +27,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_font_angle_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_font_angle_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   if ( !isParameterDoubleMatrix( valueType ) )
   {
