+++ scilab/modules/graphics/src/c/getHandleProperty/set_fractional_font_property.c
@@ -25,7 +25,7 @@
 #include "localization.h"
 
 /*------------------------------------------------------------------------*/
-int set_fractional_font_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_fractional_font_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   if (   sciGetEntityType(pobj) != SCI_SUBWIN
