+++ scilab/modules/graphics/src/c/getHandleProperty/set_grid_property.c
@@ -28,7 +28,7 @@
 #include "BasicAlgos.h"
 
 /*------------------------------------------------------------------------*/
-int set_grid_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_grid_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   int i ;
   int gridStyles[3];
