+++ scilab/modules/graphics/src/c/getHandleProperty/set_hidden_axis_color_property.c
@@ -27,7 +27,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_hidden_axis_color_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_hidden_axis_color_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   int haColor      = (int) getDoubleFromStack( stackPointer ) ;
