+++ scilab/modules/graphics/src/c/getHandleProperty/set_label_font_style_property.c
@@ -27,7 +27,7 @@
 #include "localization.h"
 
 /*------------------------------------------------------------------------*/
-int set_labels_font_style_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_labels_font_style_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   if ( !isParameterDoubleMatrix( valueType ) )
