+++ scilab/modules/graphics/src/c/getHandleProperty/set_mark_style_property.c
@@ -26,7 +26,7 @@
 #include "localization.h"
 
 /*------------------------------------------------------------------------*/
-int set_mark_style_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_mark_style_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   int status1 ;
   int status2 ;
