+++ scilab/modules/graphics/src/c/getHandleProperty/set_position_property.c
@@ -27,7 +27,7 @@
 #include "localization.h"
 #include "SetUicontrolPosition.h"
 /*------------------------------------------------------------------------*/
-int set_position_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_position_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   if(sciGetEntityType(pobj) == SCI_UICONTROL || sciGetEntityType(pobj) == SCI_FIGURE)
